+++ lib/src/data/ble/facade/ble_transport_facade_impl.dart
@@ -15,50 +15,45 @@ final class BleTransportFacadeImpl implements TransportFacade {
     required TransportSessionServer transportSessionServer,
   }) : _transportSessionClient = transportSessionClient,
        _transportSessionServer = transportSessionServer {
-    _proxyMessagesStreamController =
-        StreamController<TransportMessage>.broadcast();
-    _proxyClientMessagesStreamSubscription = _clientSession.messagesStream
-        .listen(_proxyMessagesStreamController.add);
-    _proxyServerMessagesStreamSubscription = _serverSession.messagesStream
-        .listen(_proxyMessagesStreamController.add);
+    _proxyMessagesStreamController = StreamController<TransportMessage>.broadcast();
+    _proxyClientMessagesStreamSubscription = _clientSession.messagesStream.listen(
+      _proxyMessagesStreamController.add,
+    );
+    _proxyServerMessagesStreamSubscription = _serverSession.messagesStream.listen(
+      _proxyMessagesStreamController.add,
+    );
 
-    _proxyConnectionStateStreamController =
-        StreamController<TransportSessionState>.broadcast();
-    _proxyClientConnectionStateStreamSubscription = _clientSession
-        .connectionStateStream
-        .listen(_proxyConnectionStateStreamController.add);
-    _proxyServerConnectionStateStreamSubscription = _serverSession
-        .connectionStateStream
-        .listen(_proxyConnectionStateStreamController.add);
+    _proxyConnectionStateStreamController = StreamController<TransportSessionState>.broadcast();
+    _proxyClientConnectionStateStreamSubscription = _clientSession.connectionStateStream.listen(
+      _proxyConnectionStateStreamController.add,
+    );
+    _proxyServerConnectionStateStreamSubscription = _serverSession.connectionStateStream.listen(
+      _proxyConnectionStateStreamController.add,
+    );
 
     _proxyDisconnectEventStreamController =
         StreamController<TransportSessionDisconnectEvent>.broadcast();
-    _proxyClientDisconnectEventStreamSubscription = _clientSession
-        .disconnectEventStream
-        .listen(_proxyDisconnectEventStreamController.add);
-    _proxyServerDisconnectEventStreamSubscription = _serverSession
-        .disconnectEventStream
-        .listen(_proxyDisconnectEventStreamController.add);
+    _proxyClientDisconnectEventStreamSubscription = _clientSession.disconnectEventStream.listen(
+      _proxyDisconnectEventStreamController.add,
+    );
+    _proxyServerDisconnectEventStreamSubscription = _serverSession.disconnectEventStream.listen(
+      _proxyDisconnectEventStreamController.add,
+    );
   }
 
   final TransportSessionClient _transportSessionClient;
   final TransportSessionServer _transportSessionServer;
   TransportRole _role = TransportRole.server;
 
-  TransportSession get _clientSession =>
-      _transportSessionClient as TransportSession;
+  TransportSession get _clientSession => _transportSessionClient as TransportSession;
 
-  TransportSession get _serverSession =>
-      _transportSessionServer as TransportSession;
+  TransportSession get _serverSession => _transportSessionServer as TransportSession;
 
   late final StreamController<TransportMessage> _proxyMessagesStreamController;
-  late final StreamSubscription<TransportMessage>
-  _proxyClientMessagesStreamSubscription;
-  late final StreamSubscription<TransportMessage>
-  _proxyServerMessagesStreamSubscription;
+  late final StreamSubscription<TransportMessage> _proxyClientMessagesStreamSubscription;
+  late final StreamSubscription<TransportMessage> _proxyServerMessagesStreamSubscription;
 
-  late final StreamController<TransportSessionState>
-  _proxyConnectionStateStreamController;
+  late final StreamController<TransportSessionState> _proxyConnectionStateStreamController;
   late final StreamSubscription<TransportSessionState>
   _proxyClientConnectionStateStreamSubscription;
   late final StreamSubscription<TransportSessionState>
@@ -72,8 +67,7 @@ final class BleTransportFacadeImpl implements TransportFacade {
   _proxyServerDisconnectEventStreamSubscription;
 
   @override
-  Stream<TransportMessage> get messagesStream =>
-      _proxyMessagesStreamController.stream;
+  Stream<TransportMessage> get messagesStream => _proxyMessagesStreamController.stream;
 
   @override
   Stream<TransportSessionState> get connectionStateStream =>
@@ -84,8 +78,7 @@ final class BleTransportFacadeImpl implements TransportFacade {
       _proxyDisconnectEventStreamController.stream;
 
   @override
-  Future<void> sendMessage(TransportMessage message) =>
-      transportSession.sendMessage(message);
+  Future<void> sendMessage(TransportMessage message) => transportSession.sendMessage(message);
 
   @override
   TransportSession get transportSession => switch (_role) {