+++ lib/src/data/ble/link/ble_link_base.dart
@@ -20,8 +20,7 @@ abstract base class BleLinkBase implements TransportLink {
   }
 
   @override
-  Stream<Uint8List> get incomingRawMessageStream =>
-      _incomingRawMessageController.stream;
+  Stream<Uint8List> get incomingRawMessageStream => _incomingRawMessageController.stream;
 
   @override
   Stream<void> get linkLostStream => _linkLostController.stream;
@@ -36,10 +35,7 @@ abstract base class BleLinkBase implements TransportLink {
     }
 
     final int messageId = _nextOutgoingMessageId();
-    final List<Uint8List> frames = _frameCodec.fragment(
-      data,
-      messageId: messageId,
-    );
+    final List<Uint8List> frames = _frameCodec.fragment(data, messageId: messageId);
     for (final Uint8List frame in frames) {
       await sendPhysicalFrame(frame);
     }